Revert "[Feature] Add Pagination to Tech Docs Table"
