--- .changeset/curly-beans-brake.md
@@ -1,5 +1,0 @@
----
-'@backstage/plugin-techdocs': minor
----
-
-Add pagination support to TechDocs Index Page and make it the default
