+++ packages/app/src/App.tsx
@@ -164,10 +164,7 @@ const routes = (
         />
       }
     />
-    <Route
-      path="/docs"
-      element={<TechDocsIndexPage pagination={{ mode: 'offset', limit: 20 }} />}
-    />
+    <Route path="/docs" element={<TechDocsIndexPage />} />
     <Route
       path="/docs/:namespace/:kind/:name/*"
       element={<TechDocsReaderPage />}
