+++ plugins/techdocs/src/home/components/Tables/DocsTable.tsx
@@ -18,7 +18,11 @@ import React from 'react';
 import useCopyToClipboard from 'react-use/esm/useCopyToClipboard';
 
 import { configApiRef, useApi, useRouteRef } from '@backstage/core-plugin-api';
-import { Entity } from '@backstage/catalog-model';
+import { Entity, RELATION_OWNED_BY } from '@backstage/catalog-model';
+import {
+  getEntityRelations,
+  humanizeEntityRef,
+} from '@backstage/plugin-catalog-react';
 import { rootDocsRouteRef } from '../../../routes';
 import {
   EmptyState,
@@ -29,9 +33,9 @@ import {
   TableProps,
 } from '@backstage/core-components';
 import { actionFactories } from './actions';
-import { columnFactories, defaultColumns } from './columns';
+import { columnFactories } from './columns';
+import { toLowerMaybe } from '../../../helpers';
 import { DocsTableRow } from './types';
-import { entitiesToDocsMapper } from './helpers';
 
 /**
  * Props for {@link DocsTable}.
@@ -47,6 +51,14 @@ export type DocsTableProps = {
   options?: TableOptions<DocsTableRow>;
 };
 
+const defaultColumns: TableColumn<DocsTableRow>[] = [
+  columnFactories.createTitleColumn({ hidden: true }),
+  columnFactories.createNameColumn(),
+  columnFactories.createOwnerColumn(),
+  columnFactories.createKindColumn(),
+  columnFactories.createTypeColumn(),
+];
+
 /**
  * Component which renders a table documents
  *
@@ -59,11 +71,26 @@ export const DocsTable = (props: DocsTableProps) => {
   const config = useApi(configApiRef);
   if (!entities) return null;
 
-  const documents = entitiesToDocsMapper(
-    entities,
-    getRouteToReaderPageFor,
-    config,
-  );
+  const documents = entities.map(entity => {
+    const ownedByRelations = getEntityRelations(entity, RELATION_OWNED_BY);
+    return {
+      entity,
+      resolved: {
+        docsUrl: getRouteToReaderPageFor({
+          namespace: toLowerMaybe(
+            entity.metadata.namespace ?? 'default',
+            config,
+          ),
+          kind: toLowerMaybe(entity.kind, config),
+          name: toLowerMaybe(entity.metadata.name, config),
+        }),
+        ownedByRelations,
+        ownedByRelationsTitle: ownedByRelations
+          .map(r => humanizeEntityRef(r, { defaultKind: 'group' }))
+          .join(', '),
+      },
+    };
+  });
 
   const defaultActions: TableProps<DocsTableRow>['actions'] = [
     actionFactories.createCopyDocsUrlAction(copyToClipboard),
