+++ plugins/techdocs/src/home/components/Tables/EntityListDocsTable.tsx
@@ -24,19 +24,14 @@ import {
   TableProps,
   WarningPanel,
 } from '@backstage/core-components';
-import { configApiRef, useApi, useRouteRef } from '@backstage/core-plugin-api';
 import {
   useEntityList,
   useStarredEntities,
 } from '@backstage/plugin-catalog-react';
 import { DocsTable } from './DocsTable';
-import { OffsetPaginatedDocsTable } from './OffsetPaginatedDocsTable';
-import { CursorPaginatedDocsTable } from './CursorPaginatedDocsTable';
 import { actionFactories } from './actions';
-import { columnFactories, defaultColumns } from './columns';
+import { columnFactories } from './columns';
 import { DocsTableRow } from './types';
-import { rootDocsRouteRef } from '../../../routes';
-import { entitiesToDocsMapper } from './helpers';
 
 /**
  * Props for {@link EntityListDocsTable}.
@@ -56,12 +51,9 @@ export type EntityListDocsTableProps = {
  */
 export const EntityListDocsTable = (props: EntityListDocsTableProps) => {
   const { columns, actions, options } = props;
-  const { loading, error, entities, filters, paginationMode, pageInfo } =
-    useEntityList();
+  const { loading, error, entities, filters } = useEntityList();
   const { isStarredEntity, toggleStarredEntity } = useStarredEntities();
   const [, copyToClipboard] = useCopyToClipboard();
-  const getRouteToReaderPageFor = useRouteRef(rootDocsRouteRef);
-  const config = useApi(configApiRef);
 
   const title = capitalize(filters.user?.value ?? 'all');
 
@@ -72,38 +64,6 @@ export const EntityListDocsTable = (props: EntityListDocsTableProps) => {
       toggleStarredEntity,
     ),
   ];
-
-  const documents = entitiesToDocsMapper(
-    entities,
-    getRouteToReaderPageFor,
-    config,
-  );
-
-  if (paginationMode === 'cursor') {
-    return (
-      <CursorPaginatedDocsTable
-        columns={columns || defaultColumns}
-        isLoading={loading}
-        title={title}
-        actions={actions || defaultActions}
-        options={options}
-        data={documents}
-        next={pageInfo?.next}
-        prev={pageInfo?.prev}
-      />
-    );
-  } else if (paginationMode === 'offset') {
-    return (
-      <OffsetPaginatedDocsTable
-        columns={columns || defaultColumns}
-        isLoading={loading}
-        title={title}
-        actions={actions || defaultActions}
-        options={options}
-        data={documents}
-      />
-    );
-  }
 
   if (error) {
     return (
