+++ plugins/techdocs/src/home/components/Tables/columns.tsx
@@ -85,11 +85,3 @@ export const columnFactories = {
     };
   },
 };
-
-export const defaultColumns: TableColumn<DocsTableRow>[] = [
-  columnFactories.createTitleColumn({ hidden: true }),
-  columnFactories.createNameColumn(),
-  columnFactories.createOwnerColumn(),
-  columnFactories.createKindColumn(),
-  columnFactories.createTypeColumn(),
-];
