+++ plugins/techdocs/src/home/components/TechDocsCustomHome.test.tsx
@@ -14,11 +14,7 @@
  * limitations under the License.
  */
 
-import {
-  catalogApiRef,
-  starredEntitiesApiRef,
-  MockStarredEntitiesApi,
-} from '@backstage/plugin-catalog-react';
+import { catalogApiRef } from '@backstage/plugin-catalog-react';
 import { catalogApiMock } from '@backstage/plugin-catalog-react/testUtils';
 import { renderInTestApp, TestApiRegistry } from '@backstage/test-utils';
 import { screen } from '@testing-library/react';
@@ -41,10 +37,7 @@ const mockCatalogApi = catalogApiMock({
 });
 
 describe('TechDocsCustomHome', () => {
-  const apiRegistry = TestApiRegistry.from(
-    [catalogApiRef, mockCatalogApi],
-    [starredEntitiesApiRef, new MockStarredEntitiesApi()],
-  );
+  const apiRegistry = TestApiRegistry.from([catalogApiRef, mockCatalogApi]);
 
   it('should render a TechDocs home page', async () => {
     const tabsConfig = [
