+++ plugins/techdocs/src/home/components/TechDocsCustomHome.tsx
@@ -23,7 +23,6 @@ import {
   catalogApiRef,
   CatalogApi,
   useEntityOwnership,
-  EntityListProvider,
 } from '@backstage/plugin-catalog-react';
 import { Entity } from '@backstage/catalog-model';
 import { DocsTable } from './Tables';
@@ -128,9 +127,7 @@ const CustomPanel = ({
         ) : null}
       </ContentHeader>
       <div className={classes.panelContainer}>
-        <EntityListProvider>
-          <Panel data-testid="techdocs-custom-panel" entities={shownEntities} />
-        </EntityListProvider>
+        <Panel data-testid="techdocs-custom-panel" entities={shownEntities} />
       </div>
     </>
   );
