+++ plugins/techdocs/src/home/components/TechDocsIndexPage.tsx
@@ -18,7 +18,6 @@ import React from 'react';
 import { useOutlet } from 'react-router-dom';
 import { TableColumn, TableProps } from '@backstage/core-components';
 import {
-  EntityListPagination,
   EntityOwnerPickerProps,
   UserListFilterKind,
 } from '@backstage/plugin-catalog-react';
@@ -35,7 +34,6 @@ export type TechDocsIndexPageProps = {
   columns?: TableColumn<DocsTableRow>[];
   actions?: TableProps<DocsTableRow>['actions'];
   ownerPickerMode?: EntityOwnerPickerProps['mode'];
-  pagination?: EntityListPagination;
 };
 
 export const TechDocsIndexPage = (props: TechDocsIndexPageProps) => {
